Moved ValidationTestCase to a utils.py file.
